% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qmd.R
\name{create_qmd}
\alias{create_qmd}
\title{Using \code{create_qmd}}
\usage{
create_qmd(file_name = NULL,
           ext_type = "short_report_pdf")
}
\arguments{
\item{file_name}{A string that will become the name of the Quarto report}

\item{ext_type}{Specify what type of report you want to make. Options include:
\itemize{
\item \code{short_report_pdf}: creates a Quarto doc that renders to a short, portrait, PDF report
\item \code{short_report_lscp_pdf}: creates a Quarto doc that renders to a short, landscape, PDF report
\item \code{long_report_pdf}: creates a Quarto doc that renders to a long, portrait, PDF report
\item \code{long_report_lscp_pdf}: creates a Quarto doc that renders to a long, landscape, PDF report
}}
}
\value{
This creates and opens a new Quarto doc with the name specified in \code{file_name}.
}
\description{
Use this function to create a pre-formatted Quarto doc that renders to PDF with Typst.
}
\examples{
\dontrun{
create_qmd(file_name = "new_report",
           ext_type = "short_report_pdf")
}
}
